# openapi-v1.yaml
openapi: 3.1.0
info:
  title: ARB AI Agent API
  description: Architecture Review Board AI Agent Backend
  version: 1.0.0
paths:
...
  /api/v1/metadata/steps:
    get:
      tags:
      - v1 / metadata
      summary: Get Steps
      description: Get all submission steps
      operationId: get_steps_api_v1_metadata_steps_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/metadata/domains:
    get:
      tags:
      - v1 / metadata
      summary: Get Domains
      description: Get all domains
      operationId: get_domains_api_v1_metadata_domains_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/metadata/domains/seq-number/{seq_number}:
    get:
      tags:
      - v1 / metadata
      summary: Get Domain By Seq Number
      description: Get a domain by its sequence number
      operationId: get_domain_by_seq_number_api_v1_metadata_domains_seq_number__seq_number__get
      parameters:
      - name: seq_number
        in: path
        required: true
        schema:
          type: integer
          title: Seq Number
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/metadata/domains/step/{step_id}:
    get:
      tags:
      - v1 / metadata
      summary: Get Domains For Step
      description: Get domains for a specific step
      operationId: get_domains_for_step_api_v1_metadata_domains_step__step_id__get
      parameters:
      - name: step_id
        in: path
        required: true
        schema:
          type: string
          title: Step Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/metadata/artefact-types:
    get:
      tags:
      - v1 / metadata
      summary: Get Artefact Types
      description: Get all artefact types
      operationId: get_artefact_types_api_v1_metadata_artefact_types_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/metadata/artefact-templates/domain/{domain_slug}:
    get:
      tags:
      - v1 / metadata
      summary: Get Artefact Templates
      description: Get artefact templates for a specific domain
      operationId: get_artefact_templates_api_v1_metadata_artefact_templates_domain__domain_slug__get
      parameters:
      - name: domain_slug
        in: path
        required: true
        schema:
          type: string
          title: Domain Slug
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/metadata/checklist-subsections/domain/{domain_slug}:
    get:
      tags:
      - v1 / metadata
      summary: Get Checklist Subsections
      description: Get checklist subsections for a specific domain
      operationId: get_checklist_subsections_api_v1_metadata_checklist_subsections_domain__domain_slug__get
      parameters:
      - name: domain_slug
        in: path
        required: true
        schema:
          type: string
          title: Domain Slug
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/metadata/ptx-gates:
    get:
      tags:
      - v1 / metadata
      summary: Get Ptx Gates
      description: Get all PTX gates
      operationId: get_ptx_gates_api_v1_metadata_ptx_gates_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/metadata/architecture-dispositions:
    get:
      tags:
      - v1 / metadata
      summary: Get Architecture Dispositions
      description: Get all architecture dispositions
      operationId: get_architecture_dispositions_api_v1_metadata_architecture_dispositions_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/metadata/ea-principles:
    get:
      tags:
      - v1 / metadata
      summary: Get Ea Principles
      description: Get all EA principles
      operationId: get_ea_principles_api_v1_metadata_ea_principles_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/metadata/ea-principles/domain/{domain_slug}:
    get:
      tags:
      - v1 / metadata
      summary: Get Ea Principles For Domain
      description: Get EA principles for a specific domain
      operationId: get_ea_principles_for_domain_api_v1_metadata_ea_principles_domain__domain_slug__get
      parameters:
      - name: domain_slug
        in: path
        required: true
        schema:
          type: string
          title: Domain Slug
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/metadata/form-fields/step/{step_id}:
    get:
      tags:
      - v1 / metadata
      summary: Get Form Fields
      description: Get form fields for a specific step
      operationId: get_form_fields_api_v1_metadata_form_fields_step__step_id__get
      parameters:
      - name: step_id
        in: path
        required: true
        schema:
          type: string
          title: Step Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/metadata/question-options/question/{question_id}:
    get:
      tags:
      - v1 / metadata
      summary: Get Question Options
      description: Get question options for a specific question
      operationId: get_question_options_api_v1_metadata_question_options_question__question_id__get
      parameters:
      - name: question_id
        in: path
        required: true
        schema:
          type: string
          title: Question Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/metadata/question-options:
    get:
      tags:
      - v1 / metadata
      summary: Get All Question Options
      description: Get all question options
      operationId: get_all_question_options_api_v1_metadata_question_options_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/metadata/all:
    get:
      tags:
      - v1 / metadata
      summary: Get All Metadata
      description: Get all metadata in a single call
      operationId: get_all_metadata_api_v1_metadata_all_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v2/metadata/steps:
    get:
      tags:
      - v2 / metadata
      summary: Get Steps
      description: Get all submission steps
      operationId: get_steps_api_v2_metadata_steps_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/metadata/domains:
    get:
      tags:
      - v2 / metadata
      summary: Get Domains
      description: Get all domains
      operationId: get_domains_api_v2_metadata_domains_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/metadata/domains/seq-number/{seq_number}:
    get:
      tags:
      - v2 / metadata
      summary: Get Domain By Seq Number
      description: Get a domain by its sequence number
      operationId: get_domain_by_seq_number_api_v2_metadata_domains_seq_number__seq_number__get
      parameters:
      - name: seq_number
        in: path
        required: true
        schema:
          type: integer
          title: Seq Number
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/metadata/domains/step/{step_id}:
    get:
      tags:
      - v2 / metadata
      summary: Get Domains For Step
      description: Get domains for a specific step
      operationId: get_domains_for_step_api_v2_metadata_domains_step__step_id__get
      parameters:
      - name: step_id
        in: path
        required: true
        schema:
          type: string
          title: Step Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v2/metadata/artefact-types:
    get:
      tags:
      - v2 / metadata
      summary: Get Artefact Types
      description: Get all artefact types
      operationId: get_artefact_types_api_v2_metadata_artefact_types_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/metadata/artefact-templates/domain/{domain_slug}:
    get:
      tags:
      - v2 / metadata
      summary: Get Artefact Templates
      description: Get artefact templates for a specific domain
      operationId: get_artefact_templates_api_v2_metadata_artefact_templates_domain__domain_slug__get
      parameters:
      - name: domain_slug
        in: path
        required: true
        schema:
          type: string
          title: Domain Slug
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/metadata/checklist-subsections/domain/{domain_slug}:
    get:
      tags:
      - v2 / metadata
      summary: Get Checklist Subsections
      description: Get checklist subsections for a specific domain
      operationId: get_checklist_subsections_api_v2_metadata_checklist_subsections_domain__domain_slug__get
      parameters:
      - name: domain_slug
        in: path
        required: true
        schema:
          type: string
          title: Domain Slug
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/metadata/ptx-gates:
    get:
      tags:
      - v2 / metadata
      summary: Get Ptx Gates
      description: Get all PTX gates
      operationId: get_ptx_gates_api_v2_metadata_ptx_gates_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/metadata/architecture-dispositions:
    get:
      tags:
      - v2 / metadata
      summary: Get Architecture Dispositions
      description: Get all architecture dispositions
      operationId: get_architecture_dispositions_api_v2_metadata_architecture_dispositions_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/metadata/ea-principles:
    get:
      tags:
      - v2 / metadata
      summary: Get Ea Principles
      description: Get all EA principles
      operationId: get_ea_principles_api_v2_metadata_ea_principles_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/metadata/ea-principles/domain/{domain_slug}:
    get:
      tags:
      - v2 / metadata
      summary: Get Ea Principles For Domain
      description: Get EA principles for a specific domain
      operationId: get_ea_principles_for_domain_api_v2_metadata_ea_principles_domain__domain_slug__get
      parameters:
      - name: domain_slug
        in: path
        required: true
        schema:
          type: string
          title: Domain Slug
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/metadata/form-fields/step/{step_id}:
    get:
      tags:
      - v2 / metadata
      summary: Get Form Fields
      description: Get form fields for a specific step
      operationId: get_form_fields_api_v2_metadata_form_fields_step__step_id__get
      parameters:
      - name: step_id
        in: path
        required: true
        schema:
          type: string
          title: Step Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/metadata/question-options/question/{question_id}:
    get:
      tags:
      - v2 / metadata
      summary: Get Question Options
      description: Get question options for a specific question
      operationId: get_question_options_api_v2_metadata_question_options_question__question_id__get
      parameters:
      - name: question_id
        in: path
        required: true
        schema:
          type: string
          title: Question Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/metadata/question-options:
    get:
      tags:
      - v2 / metadata
      summary: Get All Question Options
      description: Get all question options
      operationId: get_all_question_options_api_v2_metadata_question_options_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError